gen/requirements.txt
# Required packages
pyyaml>=6.0.1
jinja2>=3.1.2
requests>=2.31.0
pathlib>=1.0.1
typing>=3.7.4.3
